requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0.3
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
